ui: display TXs in address page
